requirements.txt
fastapi~=0.115.12
openai~=1.78.0
alembic==1.15.2
SQLAlchemy==2.0.40
asyncpg~=0.30.0
python-dotenv~=1.1.0